Initial Commit
What
=================
The contents of the mod have been broken down and organised in preparation
of the build tools.
Git evironment has been setup.
Readme created with original content from the forum post describing how
the mod works, what it contains and download links. Expanding on this, is a section for
developers.