Added - Build script changing version numbers
What
=================
Added - the ability for the build script to change all mentions of a
version number to the latest version number. This is intended for use
only on release builds. For this reason, the functionality is hidden
unless the global variable "RELEASE" is set to "True".
This is a temporary measure. The build script will soon undergo an
interaction modification change as stated in issue #20.

Issue ID: #6