Changed - Normalised string key names + corrected misspellings
What
=================
Normalised translation key names to follow a more sensible format, based
off the structure of the stringtable, disregarding the "sheet_" keyword
for the sake of shorter key names.
Corrected some spelling mistakes.
Found some string keys which are not being used. However, I left them in
place since Japanese and German translations were already made for them
and it would be a shame to waste that effort in case we have need for
them in the future. However, they should be removed before v1.0 release.